manga_translate_bot2/requirements.txt
python-telegram-bot>=22.6
Pillow>=10.3.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.4
easyocr>=1.7.1
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.109.1
